properly separate the voronoi search radius from the smoothing length of the particles

+++ src/hydro/Shadowswift/hydro_part.h
@@ -114,7 +114,7 @@ struct part {
   float h;
 
   /*! Particle search radius. */
-  float r;
+  double r;
 
   /*! Particle internal energy. */
   float u;
